move relay.go and reg.go to a module
