+++ term_conn/reg.go
@@ -1,4 +1,4 @@
-package main
+package term_conn
 
 import (
 	"errors"
